fix8/11 Main bus fixed--then rsrc file fix.

+++ common/src/main/java/com/kwwsyk/endinv/common/data/EndInvCodecStrategy.java
@@ -139,7 +139,7 @@ public interface EndInvCodecStrategy {
     static CompoundTag save(ItemStack stack, CompoundTag compoundTag) {
         ResourceLocation resourcelocation = BuiltInRegistries.ITEM.getKey(stack.getItem());
         compoundTag.putString("id", resourcelocation.toString());
-        compoundTag.putInt("Count", (byte)stack.getCount());
+        compoundTag.putInt("Count", stack.getCount());
         if (stack.getTag() != null) {
             compoundTag.put("tag", stack.getTag().copy());
         }
